requirements.txt
django==3.1.4
djangorestframework==3.12.2
gunicorn==20.0.4
django-heroku==0.3.1
whitenoise

